Changing vat value data type

+++ MarketplaceBetter.Database/dbo/Tables/Sales/Settings/VatRule.sql
@@ -3,7 +3,7 @@
 	[Id]			BIGINT						NOT NULL						IDENTITY,
 	[CountryFromId]	BIGINT						NOT NULL,
 	[CountryToId]	BIGINT						NOT NULL,
-	[VatValue]		INT							NOT NULL,
+	[VatValue]		FLOAT						NOT NULL,
 	[VatNumber]		NVARCHAR (255)				NOT NULL,
 	CONSTRAINT		[PK_VatRule]				PRIMARY KEY ([Id]),
 	CONSTRAINT		[FK_VatRule_CountryFrom]	FOREIGN KEY ([CountryFromId])	REFERENCES [dbo].[Country] ([Id]),
